fake-cucumber: move tests to a better place
